Fix timespan service

+++ frontend/src/services/timespan.js
@@ -15,10 +15,27 @@ const getAll = async () => {
     const response = await axios.get(timeSpanUrl)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -28,10 +45,27 @@ const getOne = async (id) => {
     const response = await axios.get(`${timeSpanUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -41,10 +75,27 @@ const create = async (newTime) => {
     const response = await axios.post(timeSpanUrl, newTime)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -54,10 +105,27 @@ const update = async (id, updatedDate) => {
     const response = await axios.put(`${timeSpanUrl}${id}/`, updatedDate)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -67,10 +135,27 @@ const del = async (id) => {
     const response = await axios.delete(`${timeSpanUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
